Create getStackDetail.py
I made this script to facilitate inventory purposes at my workplace (Palm Beach School District). This script grabs the mac addresses of a stack, along with their serial numbers and version numbers. It then sends it to the desired e-mail. This script was made in Python and can only be used in XMC.